+++ context/app/static/js/components/Detail/CollectionsAffiliationsTable/index.js
@@ -1,0 +1,3 @@
+import CollectionsAfiliationsTable from './CollectionsAffiliationsTable';
+
+export default CollectionsAfiliationsTable;