+++ context/app/static/js/components/Detail/visualization/Visualization/Visualization.jsx
@@ -1,8 +1,10 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState, useEffect, useCallback } from 'react';
 import { Vitessce } from 'vitessce';
 import Paper from '@material-ui/core/Paper';
 import Link from '@material-ui/core/Link';
 import FullscreenRoundedIcon from '@material-ui/icons/FullscreenRounded';
+import debounce from 'lodash/debounce';
+import Bowser from 'bowser';
 
 import { Alert } from 'js/shared-styles/alerts';
 import DropdownListbox from 'js/shared-styles/dropdowns/DropdownListbox';
@@ -10,6 +12,7 @@ import DropdownListboxOption from 'js/shared-styles/dropdowns/DropdownListboxOpt
 import { SecondaryBackgroundTooltip } from 'js/shared-styles/tooltips';
 import useVisualizationStore from 'js/stores/useVisualizationStore';
 import VisualizationThemeSwitch from '../VisualizationThemeSwitch';
+import VisualizationShareButton from '../VisualizationShareButton';
 import {
   vitessceFixedHeight,
   bodyExpandedCSS,
@@ -18,13 +21,22 @@ import {
   StyledHeaderText,
   StyledHeaderRight,
   ExpandButton,
-  EscSnackbar,
+  VitessceInfoSnackbar,
   ErrorSnackbar,
   ExpandableDiv,
   StyledFooterText,
   SelectionButton,
 } from './style';
+import { useVitessceConfig } from './hooks';
 import 'vitessce/dist/es/production/static/css/index.css';
+
+function sniffBrowser() {
+  const { browser } = Bowser.parse(window.navigator.userAgent);
+  return browser.name;
+}
+
+const FIREFOX_WARNING =
+  'On older versions of Firefox, zooming in on visualizations is very slow; Upgrading to the latest version will fix the problem.';
 
 const visualizationStoreSelector = (state) => ({
   vizIsFullscreen: state.vizIsFullscreen,
@@ -33,6 +45,10 @@ const visualizationStoreSelector = (state) => ({
   vizTheme: state.vizTheme,
   vizEscSnackbarIsOpen: state.vizEscSnackbarIsOpen,
   setVizEscSnackbarIsOpen: state.setVizEscSnackbarIsOpen,
+  setVitessceState: state.setVitessceState,
+  onCopyUrlWarning: state.onCopyUrlWarning,
+  onCopyUrlSnackbarOpen: state.onCopyUrlSnackbarOpen,
+  setOnCopyUrlSnackbarOpen: state.setOnCopyUrlSnackbarOpen,
 });
 
 function Visualization(props) {
@@ -45,11 +61,24 @@ function Visualization(props) {
     vizTheme,
     vizEscSnackbarIsOpen,
     setVizEscSnackbarIsOpen,
+    setVitessceState,
+    onCopyUrlWarning,
+    onCopyUrlSnackbarOpen,
+    setOnCopyUrlSnackbarOpen,
   } = useVisualizationStore(visualizationStoreSelector);
 
   const [vitessceErrors, setVitessceErrors] = useState([]);
-  const [vitessceSelection, setVitessceSelection] = useState(0);
+  const [isVisibleFirefoxWarning, setIsVisibleFirefoxWarning] = useState(sniffBrowser() === 'Firefox');
 
+  // Get the vitessce configuration from the url if available and set the selection if it is a multi-dataset.
+  const { vitessceConfig, vitessceSelection, setVitessceSelection } = useVitessceConfig({
+    vitData,
+    setVitessceState,
+    setVitessceErrors,
+  });
+
+  // The application is very slow without debouncing since state can be quite large.
+  const handleVitessceConfigDebounced = useCallback(debounce(setVitessceState, 250, { trailing: true }), []);
   function removeError(message) {
     setVitessceErrors((prev) => prev.filter((d) => d !== message));
   }
@@ -75,73 +104,101 @@ function Visualization(props) {
       window.removeEventListener('keydown', onKeydown);
     };
   }, [collapseViz]);
+  const isMultiDataset = Array.isArray(vitessceConfig);
 
   return (
-    <StyledSectionContainer id="visualization">
-      <StyledHeader>
-        <StyledHeaderText>Visualization</StyledHeaderText>
-        <StyledHeaderRight>
-          <VisualizationThemeSwitch />
-          <SecondaryBackgroundTooltip title="Switch to Fullscreen">
-            <ExpandButton size="small" onClick={expandViz} variant="contained">
-              <FullscreenRoundedIcon color="primary" />
-            </ExpandButton>
-          </SecondaryBackgroundTooltip>
-          {Array.isArray(vitData) ? (
-            <DropdownListbox
-              buttonComponent={SelectionButton}
-              optionComponent={DropdownListboxOption}
-              selectedOptionIndex={vitessceSelection}
-              options={vitData}
-              selectOnClick={setSelectionAndClearErrors}
-              getOptionLabel={(v) => v.name}
-              id="visualization-data"
-            />
-          ) : null}
-        </StyledHeaderRight>
-      </StyledHeader>
-      <Paper>
-        <ExpandableDiv $isExpanded={vizIsFullscreen} $theme={vizTheme}>
-          <EscSnackbar
-            anchorOrigin={{
-              vertical: 'top',
-              horizontal: 'center',
-            }}
-            open={vizEscSnackbarIsOpen}
-            autoHideDuration={4000}
-            onClose={() => setVizEscSnackbarIsOpen(false)}
-            message="Press [esc] to exit full window."
-          />
-          {vitessceErrors.length > 0 && (
-            <ErrorSnackbar
+    vitessceConfig &&
+    // Don't render multi-datasets unless they have a selection from the list of options in vitessceConfig.
+    (!isMultiDataset || Number.isInteger(vitessceSelection)) && (
+      <StyledSectionContainer id="visualization">
+        <StyledHeader>
+          <StyledHeaderText>Visualization</StyledHeaderText>
+          <StyledHeaderRight>
+            <VisualizationShareButton />
+            <VisualizationThemeSwitch />
+            <SecondaryBackgroundTooltip title="Switch to Fullscreen">
+              <ExpandButton size="small" onClick={expandViz} variant="contained">
+                <FullscreenRoundedIcon color="primary" />
+              </ExpandButton>
+            </SecondaryBackgroundTooltip>
+            {isMultiDataset && (
+              <DropdownListbox
+                buttonComponent={SelectionButton}
+                optionComponent={DropdownListboxOption}
+                selectedOptionIndex={vitessceSelection}
+                options={vitessceConfig}
+                selectOnClick={setSelectionAndClearErrors}
+                getOptionLabel={(v) => v.name}
+                id="visualization-data"
+              />
+            )}
+          </StyledHeaderRight>
+        </StyledHeader>
+        <Paper>
+          <ExpandableDiv $isExpanded={vizIsFullscreen} $theme={vizTheme}>
+            <VitessceInfoSnackbar
               anchorOrigin={{
-                vertical: 'bottom',
-                horizontal: 'right',
+                vertical: 'top',
+                horizontal: 'center',
               }}
-              open
-              key={vitessceErrors[0]}
-            >
-              <Alert severity="error" onClose={() => removeError(vitessceErrors[0])}>
-                {vitessceErrors[0]}
-              </Alert>
-            </ErrorSnackbar>
-          )}
-          <Vitessce
-            config={vitData[vitessceSelection] || vitData}
-            theme={vizTheme}
-            height={vizIsFullscreen ? null : vitessceFixedHeight}
-            onWarn={addError}
-          />
-        </ExpandableDiv>
-      </Paper>
-      <StyledFooterText variant="body2">
-        Powered by&nbsp;
-        <Link href="http://vitessce.io" target="_blank" rel="noreferrer">
-          Vitessce
-        </Link>
-      </StyledFooterText>
-      <style type="text/css">{vizIsFullscreen ? bodyExpandedCSS : null}</style>
-    </StyledSectionContainer>
+              open={isVisibleFirefoxWarning}
+              autoHideDuration={4000}
+              onClose={() => setIsVisibleFirefoxWarning(false)}
+              message={FIREFOX_WARNING}
+            />
+            <VitessceInfoSnackbar
+              anchorOrigin={{
+                vertical: 'top',
+                horizontal: 'center',
+              }}
+              open={vizEscSnackbarIsOpen}
+              autoHideDuration={4000}
+              onClose={() => setVizEscSnackbarIsOpen(false)}
+              message="Press [esc] to exit full window."
+            />
+            <VitessceInfoSnackbar
+              anchorOrigin={{
+                vertical: 'top',
+                horizontal: 'center',
+              }}
+              open={onCopyUrlSnackbarOpen}
+              autoHideDuration={4000}
+              $isWarning={onCopyUrlWarning}
+              onClose={() => setOnCopyUrlSnackbarOpen(false)}
+              message={`Shareable URL copied to clipboard. ${onCopyUrlWarning}`}
+            />
+            {vitessceErrors.length > 0 && (
+              <ErrorSnackbar
+                anchorOrigin={{
+                  vertical: 'bottom',
+                  horizontal: 'right',
+                }}
+                open
+                key={vitessceErrors[0]}
+              >
+                <Alert severity="error" onClose={() => removeError(vitessceErrors[0])}>
+                  {vitessceErrors[0]}
+                </Alert>
+              </ErrorSnackbar>
+            )}
+            <Vitessce
+              config={vitessceConfig[vitessceSelection] || vitessceConfig}
+              theme={vizTheme}
+              onConfigChange={handleVitessceConfigDebounced}
+              height={vizIsFullscreen ? null : vitessceFixedHeight}
+              onWarn={addError}
+            />
+          </ExpandableDiv>
+        </Paper>
+        <StyledFooterText variant="body2">
+          Powered by&nbsp;
+          <Link href="http://vitessce.io" target="_blank" rel="noreferrer">
+            Vitessce
+          </Link>
+        </StyledFooterText>
+        <style type="text/css">{vizIsFullscreen && bodyExpandedCSS}</style>
+      </StyledSectionContainer>
+    )
   );
 }