+++ context/app/static/js/components/files/FileBrowserFile/FileBrowserFile.spec.js
@@ -60,8 +60,70 @@ test('has correct left margin', () => {
     </FilesProviders>,
   );
 
-  // depth * indentation multiplier * 8px spacing unit
-  const expectedMargin = depth * 1.5 * 8;
+  // depth * indentation multiplier * 8px spacing unit + width of dir arrow icon
+  const expectedMargin = depth * 1.5 * 8 + 24;
 
   expect(screen.getByTestId('file-indented-div')).toHaveStyle(`margin-left: ${expectedMargin}px`);
+});
+
+test('displays QA chip when is_qa_qc is true', () => {
+  const fileObj = {
+    rel_path: 'fakepath',
+    edam_term: 'faketerm',
+    description: 'fakedescription',
+    file: 'fakefile',
+    is_qa_qc: true,
+    size: 1000,
+  };
+
+  const depth = 0;
+
+  render(
+    <FilesProviders>
+      <FileBrowserFile fileObj={fileObj} depth={depth} />
+    </FilesProviders>,
+  );
+
+  expect(screen.getByText('QA')).toBeInTheDocument();
+});
+
+test('does not display QA chip when is_qa_qc is not provided', () => {
+  const fileObj = {
+    rel_path: 'fakepath',
+    edam_term: 'faketerm',
+    description: 'fakedescription',
+    file: 'fakefile',
+    size: 1000,
+  };
+
+  const depth = 0;
+
+  render(
+    <FilesProviders>
+      <FileBrowserFile fileObj={fileObj} depth={depth} />
+    </FilesProviders>,
+  );
+
+  expect(screen.queryByText('QA')).toBeNull();
+});
+
+test('does not display QA chip when is_qa_qc is false', () => {
+  const fileObj = {
+    rel_path: 'fakepath',
+    edam_term: 'faketerm',
+    description: 'fakedescription',
+    file: 'fakefile',
+    size: 1000,
+    is_qa_qc: false,
+  };
+
+  const depth = 0;
+
+  render(
+    <FilesProviders>
+      <FileBrowserFile fileObj={fileObj} depth={depth} />
+    </FilesProviders>,
+  );
+
+  expect(screen.queryByText('QA')).toBeNull();
 });