+++ context/app/static/js/components/files/FileBrowserNode/FileBrowserNode.jsx
@@ -3,20 +3,13 @@ import PropTypes from 'prop-types';
 
 import FileBrowserDirectory from '../FileBrowserDirectory';
 import FileBrowserFile from '../FileBrowserFile';
-import { Column } from './style';
 
 function FileBrowserNode(props) {
   const { fileSubTree, depth } = props;
   return Object.entries(fileSubTree).map(([k, v]) => {
     // if the object contains array of files, display all files
     if (k === 'files') {
-      return (
-        <Column key={`${k}-${depth}`}>
-          {v.map((file) => (
-            <FileBrowserFile key={file.rel_path} fileObj={file} depth={depth} />
-          ))}
-        </Column>
-      );
+      return v.map((file) => <FileBrowserFile key={file.rel_path} fileObj={file} depth={depth} />);
     }
     // if the object contains additional directories, display dir and continue down
     return (