+++ context/app/static/js/pages/Collection/Collection.jsx
@@ -4,15 +4,25 @@ import PropTypes from 'prop-types';
 import { LightBlueLink } from 'js/shared-styles/Links';
 import Summary from 'js/components/Detail/Summary';
 import CollectionDatasetsTable from 'js/components/Detail/CollectionDatasetsTable';
-import CollectionCreatorsTable from 'js/components/Detail/CollectionCreatorsTable';
-
+import CollectionsAffiliationsTable from 'js/components/Detail/CollectionsAffiliationsTable';
 import useSendUUIDEvent from 'js/components/Detail/useSendUUIDEvent';
 
 function Collection(props) {
   const { collection: collectionData } = props;
-  const { uuid, entitytype, display_doi, name, description } = collectionData;
+  const {
+    uuid,
+    entity_type,
+    display_doi,
+    title,
+    description,
+    create_timestamp,
+    last_modified_timestamp,
+    contacts,
+    datasets,
+    creators,
+  } = collectionData;
 
-  useSendUUIDEvent(entitytype, collectionData);
+  useSendUUIDEvent(entity_type, collectionData);
 
   return (
     <div>
@@ -20,17 +30,20 @@ function Collection(props) {
         <>
           <Summary
             uuid={uuid}
-            entity_type={entitytype}
+            entity_type={entity_type}
             display_doi={display_doi}
-            collectionName={name}
+            collectionName={title}
             description={description}
+            create_timestamp={create_timestamp}
+            last_modified_timestamp={last_modified_timestamp}
           >
             <LightBlueLink href="https://www.doi.org" target="_blank" rel="noopener noreferrer" variant="body1">
               doi.org
             </LightBlueLink>
           </Summary>
-          {'items' in collectionData && <CollectionDatasetsTable datasets={collectionData.items} />}
-          {'creators' in collectionData && <CollectionCreatorsTable creators={collectionData.creators} />}
+          {'contacts' in collectionData && <CollectionsAffiliationsTable affiliations={contacts} title="Contacts" />}
+          {'datasets' in collectionData && <CollectionDatasetsTable datasets={datasets} />}
+          {'creators' in collectionData && <CollectionsAffiliationsTable affiliations={creators} title="Creators" />}
         </>
       )}
     </div>