+++ context/app/static/js/stores/useFilesStore.js
@@ -1,0 +1,8 @@
+import create from 'zustand';
+
+const useFilesStore = create((set) => ({
+  displayOnlyQaQc: false,
+  toggleDisplayOnlyQaQc: () => set((state) => ({ displayOnlyQaQc: !state.displayOnlyQaQc })),
+}));
+
+export default useFilesStore;