+++ context/app/markdown/CHANGELOG.md
@@ -1,3 +1,22 @@
+## v0.7.3 - 2020-11-25
+
+- Tweak NGINX settings to allow caching on client.
+- Rearrange Dockerfile to improve layer caching.
+- Compress webpack assets with gzip and enable gzip in nginx.
+- Fix conditional in test.sh
+- Try out github actions for ci.
+- Upgrade Vitessce to allow for shareable linkes, and add sharing functionality.
+
+
+## v0.7.2 - 2020-11-23
+
+- Update CCF template to use the same analytics key as the rest of the site.
+- Change fullscreen visualization icon.
+- Fix spacing between description and dates in entity detail page.
+- Return collections page.
+- Add contacts table to collection page.
+
+
 ## v0.7.1 - 2020-11-16
 
 - Fix injection order for Search stylesheet.